requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
cffi==1.14.4
cryptography==3.3.1
DAWG-Python==0.7.2
decorator==4.4.2
docopt==0.6.2
fuzzywuzzy==0.18.0
pycparser==2.20
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
python-telegram-bot==13.1
pytz==2020.4
six==1.15.0
tornado==6.1
tzlocal==2.1
